+++ src/main/java/com/vectorcalculator/DiveSolver.java
@@ -195,7 +195,7 @@ public class DiveSolver implements SolverInterface {
             maximizer = VectorCalculator.getMaximizer();
             maximizer.maximize();
             if (solveFirstDive)
-                maximizer.isDiveCapBouncePossible(-1, singleThrowAllowed, false, mcctAllowed, !singleThrowAllowed && ttAllowed != TripleThrow.YES, p.tripleThrow != TripleThrow.NO); 
+                maximizer.isDiveCapBouncePossible(-1, singleThrowAllowed, false, mcctAllowed, !singleThrowAllowed && ttAllowed != TripleThrow.YES, ttAllowed != TripleThrow.NO); 
         }
 
         if (!solveFirstDive && !solveSecondDive) {
@@ -256,7 +256,7 @@ public class DiveSolver implements SolverInterface {
         p.diveFirstFrameDecel = 0;
         maximizer.edgeCBAngleIncrement = edgeCBAngleIncrement;
         bestDisp = maximizer.maximize();
-        int ctType = maximizer.isDiveCapBouncePossible(-1, singleThrowAllowed, false, mcctAllowed, !singleThrowAllowed && ttAllowed != TripleThrow.YES, p.tripleThrow != TripleThrow.NO);
+        int ctType = maximizer.isDiveCapBouncePossible(-1, singleThrowAllowed, false, mcctAllowed, !singleThrowAllowed && ttAllowed != TripleThrow.YES, ttAllowed != TripleThrow.NO);
         return ctType;
     }